Add --skip-bad-hosts-in-list option
The idea of this option is to cover a use-case in between the default
and the full --skip-bad-hosts. In this use-case, we have just one or
two servers that we expect might be down, but we still want to fail fast
in the case that some other server is unexpectedly down.

For example, if we have

    env.roledefs['gates'] = ['gate1', 'gate2', 'gate3', 'gate4']

and we need to re-deploy configuration to all the gates when one of
them has an error, then it would be convenient to write (if say
`gate2` has gone down)

    fab -R gates --skip-bad-hosts-in-list=gate2 deploy_config

or more pointedly

    @task
    def disable_gate():
        with settings(parallel=True, skip_bad_hosts_in_list=[env.host]):
            execute(deploy_config, role='gates')

    ... fab -H gate2 disable_gate ...

This is almost the same idea as `--exclude-hosts`, except that
- fab will still try the downed server every time
- the setting is propagated through `execute()`, while `exclude_hosts` is not

It would be convenient to provide an `env.permanently_exclude_hosts` that
would be propagated through `execute()`; in fact that might be better
for this use-case.